Move thesaurus types into the /types dir.
