+++ app/api/thesauris/dictionariesModel.ts
@@ -1,7 +1,7 @@
 /** @format */
 import { instanceModel } from 'api/odm';
 import mongoose from 'mongoose';
-import { ThesaurusSchema } from 'shared/thesaurusType';
+import { ThesaurusSchema } from 'shared/types/thesaurusType';
 
 const mongoSchema = new mongoose.Schema({
   name: mongoose.Schema.Types.Mixed,
