+++ app/shared/types/thesaurusType.d.ts
